+++ CHANGELOG.md
@@ -3,6 +3,7 @@ Version History
 
 ### Changes in v1.1.0:
 
+-   Added `RTLightmap` filter optimized for lightmaps
 -   Added `hdrScale` filter parameter for manually specifying the mapping
     of HDR color values to luminance levels